@@ -38,7 +38,6 @@ import org.darisadesigns.polyglotlina.Nodes.EtyExternalParent;
 import org.darisadesigns.polyglotlina.Nodes.TypeNode;
 import org.darisadesigns.polyglotlina.Nodes.WordClassValue;
 import org.darisadesigns.polyglotlina.Nodes.WordClass;
-import org.darisadesigns.polyglotlina.PFontHandler;
 import org.darisadesigns.polyglotlina.WebInterface;
 import java.awt.Color;
 import java.awt.Component;
@@ -249,10 +248,6 @@ public final class ScrLexicon extends PFrame {
             curPopulating = localPopulating;
             forceUpdate = false;
             populateProperties();
-            ((PTextField) txtConWord).setCore(_core);
-            ((PTextField)txtLocalWord).setCore(_core);
-            ((PTextField)txtProc).setCore(_core);
-            ((PTextPane)txtDefinition).setCore(_core);
             
             // ensures multiple logograph screens can't be open at once
             btnLogographs.setEnabled(btnLogoShouldEnable());
@@ -299,8 +294,7 @@ public final class ScrLexicon extends PFrame {
                     JComponent component = classPropMap.get(curProp.getKey());
 
                     try {
-                        if (component instanceof JComboBox) {
-                            JComboBox combo = (JComboBox) component;
+                        if (component instanceof JComboBox combo) {
                             combo.setSelectedItem(((WordClass) core.getWordClassCollection()
                                     .getNodeById(curProp.getKey())).getValueById(curProp.getValue()));
                         }
@@ -317,8 +311,7 @@ public final class ScrLexicon extends PFrame {
                     JComponent component = classPropMap.get(curProp.getKey());
 
                     try {
-                        if (component instanceof PTextField) {
-                            PTextField textField = (PTextField) component;
+                        if (component instanceof PTextField textField) {
                             textField.setText(curProp.getValue());
                         }
                     } catch (Exception e) {
@@ -434,8 +427,7 @@ public final class ScrLexicon extends PFrame {
                         return;
                     }
                     ConWord curWord1 = getCurrentWord();
-                    if (classBox.getSelectedItem() instanceof WordClassValue) {
-                        WordClassValue curValue = (WordClassValue) classBox.getSelectedItem();
+                    if (classBox.getSelectedItem() instanceof WordClassValue curValue) {
                         curWord1.setClassValue(classId, curValue.getId());
                     } else {
                         // if not an instance of a value, then it's the default selection: remove class from word
@@ -652,12 +644,12 @@ public final class ScrLexicon extends PFrame {
             return;
         }
 
-        int filterType;
+        int posFilter;
 
         if (cmbTypeSrc.getValue().equals(defTypeValue)) {
-            filterType = 0;
+            posFilter = 0;
         } else {
-            filterType = ((TypeNode) cmbTypeSrc.getValue()).getId();
+            posFilter = ((TypeNode) cmbTypeSrc.getValue()).getId();
         }
         
         saveValuesTo(getCurrentWord());
@@ -688,7 +680,7 @@ public final class ScrLexicon extends PFrame {
         filter.setValue(txtConSrc.getText().trim());
         filter.setDefinition(txtDefSrc.getText().trim());
         filter.setLocalWord(txtLocalSrc.getText().trim());
-        filter.setWordTypeId(filterType);
+        filter.setWordTypeId(posFilter);
         filter.setPronunciation(txtProcSrc.getText().trim());
         filter.setFilterEtyParent(cmbRootSrc.getValue());
 
@@ -834,8 +826,7 @@ public final class ScrLexicon extends PFrame {
 
             txtErrorBox.setText(txtErrorBox.getText() + message);
             element.setBackground(hColor);
-            if (element instanceof PComboBox) {
-                PComboBox eleComb = (PComboBox) element;
+            if (element instanceof PComboBox eleComb) {
                 eleComb.makeFlash(hColor, false);
             }
 
@@ -881,15 +872,16 @@ public final class ScrLexicon extends PFrame {
         GridPane grid = new GridPane();
         javafx.scene.text.Font font = ((DesktopPropertiesManager)core.getPropertiesManager()).getFXLocalFont();
         javafx.scene.text.Font conFont = ((DesktopPropertiesManager)core.getPropertiesManager()).getFXConFont();
+        javafx.scene.text.Font menuFont = PolyGlot.getPolyGlot().getOptionsManager().getFXMenuFont();
         
         gridTitlePane = new TitledPane();
-        gridTitlePane.setFont(font);
+        gridTitlePane.setFont(menuFont);
 
         grid.setPrefWidth(4000);
         txtConSrc = new TextField();
         txtConSrc.setPromptText("Search ConWord...");
         txtConSrc.setFont(conFont);
-        txtConSrc.setTooltip(new Tooltip("Filter lexicon entries based on the value of your constructed words"));
+        txtConSrc.setTooltip(new Tooltip("Filter lexicon entries based on the value of your constructed words (filter includes conjugated word forms)"));
         txtLocalSrc = new TextField();
         txtLocalSrc.setPromptText("Search NatLang Word...");
         txtLocalSrc.setFont(font);
@@ -933,24 +925,39 @@ public final class ScrLexicon extends PFrame {
 
         grid.setVgap(4);
         grid.setPadding(new Insets(5, 5, 5, 5));
-        grid.add(new Label("Con Word: "), 0, 0);
+        Label conWordLabel = new Label("Con Word:");
+        conWordLabel.setFont(menuFont);
+        grid.add(conWordLabel, 0, 0);
         grid.add(txtConSrc, 1, 0);
-        grid.add(new Label("Local Word: "), 0, 1);
+        Label localWordLabel = new Label("Local Word: ");
+        localWordLabel.setFont(menuFont);
+        grid.add(localWordLabel, 0, 1);
         grid.add(txtLocalSrc, 1, 1);
-        grid.add(new Label("Part of Speech: "), 0, 2);
+        Label posLabel = new Label("Part of Speech: ");
+        posLabel.setFont(menuFont);
+        grid.add(posLabel, 0, 2);
         grid.add(cmbTypeSrc, 1, 2);
         grid.setPadding(new Insets(5, 5, 5, 5));
         grid.add(new Label("            "), 2, 0); // adds spacing
-        grid.add(new Label("Pronunciation: "), 3, 0);
+        Label pronunciationLabel = new Label("Pronunciation: ");
+        pronunciationLabel.setFont(menuFont);
+        grid.add(pronunciationLabel, 3, 0);
         grid.add(txtProcSrc, 4, 0);
-        grid.add(new Label("Definition: "), 3, 1);
+        Label definitionLabel = new Label("Definition: ");
+        definitionLabel.setFont(menuFont);
+        grid.add(definitionLabel, 3, 1);
         grid.add(txtDefSrc, 4, 1);
-        grid.add(new Label("Root: "), 3, 2);
+        Label rootLabel = new Label("Root: ");
+        rootLabel.setFont(menuFont);
+        grid.add(rootLabel, 3, 2);
         grid.add(cmbRootSrc, 4, 2);
-        grid.add(new Label("Illegals"), 0, 3);
+        Label illeggalsLabel = new Label("Illegals");
+        illeggalsLabel.setFont(menuFont);
+        grid.add(illeggalsLabel, 0, 3);
         grid.add(chkFindBad, 1, 3);
 
         javafx.scene.control.Button srcButton = new javafx.scene.control.Button("Filter");
+        srcButton.setFont(menuFont);
         srcButton.setOnAction((javafx.event.ActionEvent t) -> {
             runFilter();
         });
@@ -958,6 +965,7 @@ public final class ScrLexicon extends PFrame {
         
         // sets up button to clear filter
         javafx.scene.control.Button clearButton = new javafx.scene.control.Button("Clear Filter");
+        clearButton.setFont(menuFont);
         clearButton.setOnAction((javafx.event.ActionEvent t) -> {
             clearFilterInternal();
             runFilter();
@@ -1235,22 +1243,6 @@ public final class ScrLexicon extends PFrame {
     }
 
     /**
-     * Adds appropriate listeners to conword property fields
-     *
-     * @param field field to add lister to
-     * @param defValue default string value
-     */
-    private void addPropertyListeners(JComponent field, final String defValue) {
-        if (field instanceof JComboBox) {
-            final JComboBox cmbField = (JComboBox) field;
-            cmbField.addActionListener((java.awt.event.ActionEvent evt) -> {
-                setGreyFields(cmbField, defValue);
-                setWordLegality();
-            });
-        }
-    }
-
-    /**
      * Adds appropriate listeners to filter fields (java FX Control objects)
      *
      * @param field field to add listener to
@@ -1390,21 +1382,6 @@ public final class ScrLexicon extends PFrame {
     }
 
     /**
-     * Sets appropriate fields grey
-     */
-    private void setGreyFields(JComponent comp, String defValue) {
-        if (comp instanceof JComboBox) {
-            JComboBox compCmb = (JComboBox) comp;
-            if (compCmb.getSelectedItem() != null
-                    && compCmb.getSelectedItem().toString().equals(defValue)) {
-                compCmb.setForeground(Color.red);
-            } else {
-                compCmb.setForeground(Color.black);
-            }
-        }
-    }
-
-    /**
      * populates lexicon list with all words from core
      */
     private void populateLexicon() {
@@ -1505,13 +1482,10 @@ public final class ScrLexicon extends PFrame {
 
         // save all class values
         classPropMap.entrySet().forEach((entry) -> {
-            if (entry.getValue() instanceof PTextField) {
-                PTextField textField = (PTextField) entry.getValue();
+            if (entry.getValue() instanceof PTextField textField) {
                 saveWord.setClassTextValue(entry.getKey(), textField.getText());
-            } else if (entry.getValue() instanceof PComboBox) {
-                PComboBox comboBox = (PComboBox) entry.getValue();
-                if (comboBox.getSelectedItem() instanceof WordClassValue) {
-                    WordClassValue curValue = (WordClassValue) comboBox.getSelectedItem();
+            } else if (entry.getValue() instanceof PComboBox comboBox) {
+                if (comboBox.getSelectedItem() instanceof WordClassValue curValue) {
                     saveWord.setClassValue(entry.getKey(), curValue.getId());
                 } else {
                     // if not an instance of a value, then it's the default selection: remove class from word
