@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2020, Draque Thompson, draquemail@gmail.com
+ * Copyright (c) 2016-2021, Draque Thompson, draquemail@gmail.com
  * All rights reserved.
  *
  * Licensed under: MIT Licence
@@ -44,6 +44,7 @@ import javax.swing.JMenuItem;
 import javax.swing.JPopupMenu;
 import javax.swing.JTextPane;
 import javax.swing.SwingWorker;
+import org.darisadesigns.polyglotlina.CustomControls.CoreUpdateSubscriptionInterface;
 import org.darisadesigns.polyglotlina.Desktop.DesktopPropertiesManager;
 import org.darisadesigns.polyglotlina.Desktop.PolyGlot;
 
@@ -51,7 +52,7 @@ import org.darisadesigns.polyglotlina.Desktop.PolyGlot;
  *
  * @author draque
  */
-public final class PTextPane extends JTextPane {
+public final class PTextPane extends JTextPane implements CoreUpdateSubscriptionInterface {
 
     private SwingWorker worker = null;
     private final String defText;
@@ -60,15 +61,14 @@ public final class PTextPane extends JTextPane {
     private final boolean overrideFont;
 
     public PTextPane(DictCore _core, boolean _overrideFont, String _defText) {
-        core = _core;
-        defText = _defText;
         overrideFont = _overrideFont;
+        setCore(_core);
+        defText = _defText;
         this.setContentType("text/html");
         setupRightClickMenu();
 
         setupListeners();
         setText(defText);
-        setFontFromCore();
         
         this.setEditorKit(new PHTMLEditorKit());
     }
@@ -86,8 +86,7 @@ public final class PTextPane extends JTextPane {
      */
     public void setFontFromCore() {
         if (overrideFont) {
-            setFont(((DesktopPropertiesManager)core.getPropertiesManager()).getFontLocal()
-                    .deriveFont((float)PolyGlot.getPolyGlot().getOptionsManager().getMenuFontSize()));
+            setFont(((DesktopPropertiesManager)core.getPropertiesManager()).getFontLocal());
         } else {
             setFont(((DesktopPropertiesManager)core.getPropertiesManager()).getFontCon());
         }
@@ -269,7 +268,12 @@ public final class PTextPane extends JTextPane {
         });
     }
     
+    @Override
     public void setCore(DictCore _core) {
+        if (core != _core && _core != null) {
+            _core.subscribe(this);
+        }
+        
         core = _core;
         setFontFromCore();
     }
@@ -459,5 +463,10 @@ public final class PTextPane extends JTextPane {
      */
     private String getSuperText() {
         return super.getText().replaceAll(PGTUtil.RTL_CHARACTER, "").replaceAll(PGTUtil.LTR_MARKER, "");
+    }
+
+    @Override
+    public void updateFromCore() {
+        setFontFromCore();
     }
 }
