@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014-2020, Draque Thompson, draquemail@gmail.com
+ * Copyright (c) 2014-2021, Draque Thompson, draquemail@gmail.com
  * All rights reserved.
  *
  * Licensed under: MIT Licence
@@ -54,6 +54,9 @@ public class PGTUtil {
     public static final String BUILD_DATE_TIME_LOCATION = "/assets/org/DarisaDesigns/buildDate";
     public static final String[] SWADESH_LISTS = {"Original_Swadesh", "Modern_Swadesh"};
     public static final int WINDOWS_CLIPBOARD_DELAY = 15;
+    public static final int SECONDS_BETWEEN_AUTO_SAVES = 300000; // 5 minutes in microsecnds
+    public static final String AUTO_SAVE_FILE_NAME = ".pgtAutoSave.bak";
+    public static final String TROUBLE_TICKET_URL = "https://github.com/DraqueT/PolyGlot/issues/new";
 
     // properties on words
     public static final String LEXICON_XID = "lexicon";
@@ -321,7 +324,8 @@ public class PGTUtil {
     public static final int CHECKBOX_ROUNDING = 3;
 
     // UI Elements to set on OSX (copy/paste/cut)
-    public static final String[] INPUT_MAPS = {"Button.focusInputMap",
+    public static final String[] INPUT_MAPS = {
+        "Button.focusInputMap",
         "CheckBox.focusInputMap",
         "ComboBox.ancestorInputMap",
         "EditorPane.focusInputMap",
@@ -346,7 +350,8 @@ public class PGTUtil {
         "TextField.focusInputMap",
         "TextPane.focusInputMap",
         "ToggleButton.focusInputMap",
-        "Tree.focusInputMap"};
+        "Tree.focusInputMap"
+    };
 
     public static final boolean IS_OSX;
     public static final boolean IS_WINDOWS;
